# Conflicts:
#	source/neural_network/AbsNet.cpp
#	source/neural_network/INTNetwork.cpp

+++ source/neural_network/AbsNet.cpp
@@ -109,8 +109,8 @@ AbsNet *AbsNet::addPool2D(const int *kernel_size, Pooling pooling_algorithm, Pad
     } else {
         pool_out_shape = {in_shape[0],
                           in_shape[1],
-                          ceil((in_shape[2]-kernel_size[0]+1)/(double)(pool_strides[0])),
-                          ceil((in_shape[3]-kernel_size[1]+1)/(double)(pool_strides[1]))};
+                          ceil((in_shape[2]-kernel_size[0]+1)/(float)(pool_strides[0])),
+                          ceil((in_shape[3]-kernel_size[1]+1)/(float)(pool_strides[1]))};
         pool_padding = {0, 0};
     }
     std::cout << "Initialized pool dimensions" << std::endl;
